requirements.txt
argon2-cffi==18.1.0
bcrypt==3.1.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
Cython==0.28.2
dj-database-url==0.5.0
django-heroku==0.3.1
docutils==0.14
future==0.16.0
geoip2==2.9.0
gunicorn==19.8.1
idna==2.6
Jinja2==2.10
Kivy-Garden==0.1.4
MarkupSafe==1.0
maxminddb==1.4.0
numpy==1.14.3
paramiko==2.4.2
Pillow==5.1.0
pipenv==2018.5.18
psycopg2==2.7.4
pycparser==2.18
Pygments==2.2.0
pylibmc==1.5.2
python-memcached==1.59
pytz==2018.4
PyYAML==3.12
requests==2.18.4
selenium==3.12.0
six==1.11.0
sqlparse==0.2.4
tblib==1.3.2
urllib3==1.22
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
